styles

+++ db/migrate/20180417131010_create_incomes.rb
@@ -4,7 +4,7 @@ class CreateIncomes < ActiveRecord::Migration[5.1]
       t.string :acceptor
       t.string :title
       t.integer :service
-      t.decimal :amount
+      t.float :amount
       t.integer :client
       t.text :comment
       t.belongs_to :cash_box,index: true
