+++ db/migrate/20180417131046_create_consumptions.rb
@@ -2,7 +2,7 @@ class CreateConsumptions < ActiveRecord::Migration[5.1]
   def change
     create_table :consumptions do |t|
       t.string :title
-      t.decimal :amount
+      t.float :amount
       t.text :comment
       t.belongs_to :cash_box,index: true
       t.timestamps
